refactor: UserFilterListener 분리
[#695]
