+++ frontend/src/components/google-maps/map/CarFfeineMap.tsx
@@ -1,11 +1,4 @@
-import { useQueryClient } from '@tanstack/react-query';
-
 import StationMarkersContainer from '@marker/StationMarkersContainer';
-
-import { serverStationFilterAction } from '@stores/station-filters/serverStationFiltersStore';
-
-import { useCarFilters } from '@hooks/tanstack-query/station-filters/useCarFilters';
-import { useMemberFilters } from '@hooks/tanstack-query/station-filters/useMemberFilters';
 
 import ToastContainer from '@common/Toast/ToastContainer';
 
@@ -15,9 +8,8 @@ import ModalContainer from '@ui/ModalContainer';
 import Navigator from '@ui/Navigator';
 import WarningModalContainer from '@ui/WarningModalContainer';
 
-import { QUERY_KEY_STATION_MARKERS } from '@constants/queryKeys';
-
 import CarFfeineMapListener from './CarFfeineListener';
+import UserFilterListener from './UserFilterListener';
 
 const CarFfeineMap = () => {
   return (
@@ -36,21 +28,6 @@ const CarFfeineMap = () => {
       <StationMarkersContainer />
     </>
   );
-};
-
-const UserFilterListener = () => {
-  const queryClient = useQueryClient();
-  const { data: memberFilters } = useMemberFilters();
-  const { data: carFilters } = useCarFilters();
-  const { setAllServerStationFilters } = serverStationFilterAction;
-
-  if (memberFilters !== undefined) {
-    setAllServerStationFilters(memberFilters);
-    setAllServerStationFilters(carFilters);
-    queryClient.invalidateQueries([{ queryKey: [QUERY_KEY_STATION_MARKERS] }]);
-  }
-
-  return <></>;
 };
 
 export default CarFfeineMap;
